CityPulseAgent/reflex_agent/requirements.txt
flask==2.3.2
google-generativeai==0.3.2
firebase-admin==6.2.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1